PCX-2738: move event utils to package
