+++ example-shop/src/main/java/com/payoneer/checkout/exampleshop/util/Event.kt
@@ -1,14 +1,14 @@
 /*
  *
- * Copyright (c) 2022 Payoneer Germany GmbH
- * https://www.payoneer.com
- *
- * This file is open source and available under the MIT license.
- *See the LICENSE file for more information.
- *
+ *  * Copyright (c) 2022 Payoneer Germany GmbH
+ *  * https://www.payoneer.com
+ *  *
+ *  * This file is open source and available under the MIT license.
+ *  * See the LICENSE file for more information.
+ *  *
  */
 
-package com.payoneer.checkout.exampleshop
+package com.payoneer.checkout.exampleshop.util
 
 import androidx.lifecycle.MutableLiveData
